:art: Move out manager-impl to its own file
